YAMC (Yet Another Massive Commit)
1.  Selections are quasi-working now.  You can build a selection by
using the shift key and the left or right arrow key.  Or, you can
drag a selection using the mouse.

2.  Everything still flickers like mad.  In fact, it's worse now that
selections are working.

3.  Fixed UGLY bug in the code which handles deletes.

4.  Added invertRect() to the DG_Graphics API.  I have added stubs
for UNIX and PostScript.  The UNIX version needs to be replaced so that
selections can work.  The PostScript version should remain an assert,
since this method should not be called.

5.  The hit test code is finally working to my initial satisfaction.

6.  My previous attempt at fixing the linebreaker didn't work for all
cases.  If you see some really weird line breaks occuring, don't worry
about it too much.  A fix is pending.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@156 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6