1.  Changed the default point size to 10pt, to match the behavior of MS Word.  Let me know if this causes problems for anyone.
2.  Text import now works like other word processors, by interpreting
an EOL as a paragraph break.  A user convinced me to do it this way,
even though I was the advocate for the previous behavior.

3.  Widow-orphan control is now supported, with a default setting of
2 for both.

4.  On Windows, the find dialog now initializes with the focus in the
Find string textfield.

5.  An Undo of a globbed operation now properly updates the frame
window title to remove the asterisk if the document is no longer dirty.

6.  fixed bug 72 -- my previous attempted fix missed a couple of cases.

7.  fixed bug 148 -- added notifyListeners(AV_CHG_EMPTYSEL) to cmdSelect
in fv_View, which catches the double-click to select a word case.

8.  fixed bug 169 -- the updateScreen code should never have been added
to _deleteSelection.  It's gone now.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@881 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6