The file filters got worked around so that they take and honor a file type on construction for both importers and exporters. abi/src/wp/impexp/xp/ie_types.h was updated to define the file types we know of, and the filters now respond to queries about the formats they support.
Upon construction an importer or exporter, an IEFileType must
be specified (previously, only exporters required this as
importers alwyas triggered an automatic extension lookup).
If that IEFileType is IEFT_Unknown, the same old extension
lookup is attempted.  Right now, the only place an importer
is constructed is in abi/src/wp/ptbl/xp/pd_Document.cpp, and
I've temporarily changed it so that it always passes type
IEFT_Unknown.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@1350 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6