Added system.profile to win32 install manifest.
Added UseEnvLocale switch to preferences mechanism
to allow us to import POSIX LOCALE settings in the
environment and override the various LabelSet vars
in the profile.  I added the code to parse the
profile and to test the value of UseEnvLocale.

TODO actually use the various POSIX variables and
override the internal settings -- I didn't have info
on the various POSIX convensions at the time.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@1909 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6