Added PreviewWidget support for Win32 version.
Updated Win32-specific parts of Zoom dialog to make
use of new PreviewWidget.

Finished inter-field interactions of Win32 Zoom dialog
(so that spin-box is enabled only when RadioPercent is
checked, and so that PreviewWidget updates when any other
field is changed, and etc).

Added range checking on Editbox and SpinBox.

===============================================

The Win32PreviewWidget forms the base of a hierarchy
of preview widgets; this, most basic one, only has
support for WM_PAINT messages (no user input).  It is
used by the Zoom dialog, but is independent of it.
New dialogs should be able to make use of it by just
referencing it -- see the bottom of _onInitDialog()
in the Zoom dialog.

===============================================

TODO All of the inter-field interaction is highly
platform specific and likely to be different between
the various platforms (and very tedious) -- we may
want to stop and review how we are doing this before
we proceed to many more dialogs.  Part of this is
due to how the platform notification events come in
(such as how the spin-button affects the edit-box),
but part of this is our problem -- such how we
enable/disable the edit-box depending upon the current
radio-button choice and whether or not we replace the
edit-box content when a fixed-value radio-button is
selected.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@1956 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6