On UNIX, an OS that allows you to name files what you want to name them, don't mangle the filenames that The User inputs to include yucky, useless suffixes or anything else. I could rant on for hours on why this is the Correct Behavior, but I will not to avoid sparking a flamewar. Anyway, files seem to work fine now without a .abw extension, so why forcefully append one? If the answer is to be like windows, well, why don't we stick a taskbar at the bottom of the screen to make abiword on unix look more like abiword on windows. My point is that different platforms work differently, and it is a bad idea to write programs that work in exactly the same way on different platforms, disregarding the traditions and standards of each platform. Make it be like the other applications that the user is using.
Furthermore, this change brings the UNIX and Win32 versions of Abi even closer together. On Win32, a filename suffix is meant as internal data for the OS to determine which filetype it is (this is because Win32 is dumb). This is why Win32 does not show filename extensions by default. So, if a user on Windows saves a file as MyFile, then when looking at the directory with the OS (explorer) they will see a file named MyFile. On UNIX before this change, if you were to type ls after saving a file MyFile, you would see MyFile.abw. This is not the same behavior as on windows! Filename extensions are a type of metadata that win32 uses, and it is incorrect to write this metadata on platforms that do not understand it and ignore it.

I hope I've made my point well, and that this change does not meet up with a lot of resistance. As you can probably tell, I have a very strong hate for filename extensions, originally being a mac user and programmer (heh, I was the one that coded support for determining the file type with file(1) in GNOME's file manager). If there is someone who opposes this change, I would like to discuss it with you, and I will try to listen to what you have to say.

Well, at least I didn't rant on for hours as I promissed not to :).


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@2739 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6