Beta version of Lists feature including a modeless dialog for Unix. The guts of the code comes of Luke Jordan. I managed to fix some of the bugs in his. There are still many bugs in the lists code however this patch fixes bug 909 and may also fix bugs 907 and 908. The patch fixes the insert symbol bug where inserting a symbol of different font wasn't gettng passed into the document if text was selected.
The list code will only be invoked in debug builds since we use fields for list
labels. Bullets don't work at all right now. We really need better unicode
capabilities for this and/or to fix bug 924 so that standard symbols display on
Linux and Windows. Overwrite mode destroys this code completely. I will
rework the GUI a bit more but really we need the new Fields framework and
Unicode to finish Lists off.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@2831 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6