Changed the gnome toolbars.  Now, when it parses a toolbar separator, it generates a new toolbar instead.  Yeah, I know, I know... it's nothing but a hack, but it works (more or less...).  From a user point of view, the separators can now be dragged off.
In <= 800x600 the toolbar layout sucks, but It's Not My Fault (TM).  A
new layout has to be discused with the dev-list (when I will have a
valid email account...).

PS.: The only concerned file is ev_UnixGnomeToolbar.{cpp,h}.  The rest
ones have only indentation changes and such (I hope... :)


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@2895 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6