From martin.
This patch enables rtf export past fields. It fixes Ben's bug when typing
over a fully selected list. It adds robustness to multi-view situations.
It enables import of footer.abw but Abi crashes if you try to exit after
importing it. This is a marginal improvement on crashing upon trying to
import footer.abw.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/branches/release-version-0-7-12@3517 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6