(Frodo) Autoconf enhancements regarding Gnome
I reshuffled the gnome detection parts.

This resolves bug 1453 (complaints about gnome-config even in non-gnome builds)
and (most of) 1449 (libgal not detected).

All detection is currently based on gnome-detect; there is no check for
include files and libraries. I think it might be easier to do that through
the already available ac-helpers scripts; but there is probably a reason
they are not used at the moment, and I don't know it.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@5032 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6