(Frodo) Support for autoconf-2.50 and updated build information
Regrettably, the configure.in syntax of autoconf-2.13 and autoconf-2.50
are not quite compatible. As a temporary solution, I created two configure.in
files; the right one should automatically be selected by autogen.sh. This
is an awful hack that should go away ASAP.

The BUILD file was quite out of date; I have updated it (though the original
text is still included). Please read it.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@5069 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6