Backport to STABLE.  I fully intend several things with the upcoming 1.0.3. 1) Plugins will be bundled.  KJD set a great example with his bundles, we're    gonna implement at packager level.  So far just IE and TOOLS. 2) Instead of 2 gtk and 2 gnome (with and without bidi), we should have one    of each, bidi-enabled. 3) If we can build the application on a given architecture, we should be able    to build the rpms on that architecture.
So far I have encountered no argument at all to any of these three things.
They apply to the builds we upload to SourceForge (and presumably to Savannah,
unless Sc0tt wants it differently).
Note that if a plugin either doesn't build or is useless (freetrans,eg,festvox)
it got removed from this version of the spec.  Naturally some are to be put back when (re)written.
Note that if I hardwired a build option (like debug,jpeg,opt,bidi) it means that this should be used for release uploads.

I'm fairly certain that this incarnation of the spec doesn't work out-of-box, but I can't do any more with it right now not having a working copy of rpm.  I'm sure uwog and Sc0tt will fix it up all pretty and nice-like.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/branches/ABI-1-0-0-STABLE@8202 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6