UT_go_file_open(): fix when loading a fd URI (eg: fd://0 for stdin). before, this would just hang
Rest: Reworked the image importer class to only use libgsf. This will load the URI once, and then work from there. I've changed the XP and UNIX bits, but Win32 should be updated. If not, it's no big harm - the old API is still there, and there shouldn't be any regressions. Opening remote images is much faster now, as it no longer has to load the image twice (once for file type detection, once to import the image).

I'll fix the plugins next.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@19309 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6