Allow listeners to bind to frame events;
Rationale: if a document is loaded in a frame outside the control of the 
'main controller (like the abiwidget)', the controller has no way at
all to find out that the view he is listening to has been deleted; it just
continues to listen to a stale pointer. 
This commit makes the AP_Frame signal whenever a document or view has been 
replaced.



git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@21951 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6