Backport: ------------------------------------------------------------------------ r22808 | dom | 2008-02-09 23:20:50 +0100 (Sat, 09 Feb 2008) | 6 lines
More cairo work. Remove the FontConfig dependency from 
GR_UnixPangoGraphics::getCoverage(), since we can get the coverage more 
generally from the PangoFont. The "Insert Symbols" dialog behaves 
identically as before.
------------------------------------------------------------------------

Additional comment from me: I think the comment in the code about 
Times New Roman not having coverage for the " char was wrong. 
This character *does* have coverage according to Pango in 
Times New Roman, but NOT in the font that is selected by
default in the Insert Symbol dialog: Standard Symbols L. 

Double check with OOo: OOo doesn't show a nice " char either for this font,
but instead displays an inverted A. I like our behavior better of just hiding
the characters with no proper coverage in a font. Feel free to disagree.

This fixes bug 11731: AbiWord crashes on "Insert Symbol" on stable as well.



git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/branches/ABI-2-6-0-STABLE@25064 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6