More progress bar goodness. Implement a "pulsing" progress bar for those long running progresses where we do not know how much more there is to go.
We set this by setting the progress type to "PROGRESS_INDEFINATE" as shown 
below:
	pStatusBar->setStatusProgressType(0,100,PROGRESS_INDEFINATE);

These are useful during the first phase of document import where we fill the 
piecetable.

The is also a new translatable string. "Inserted paragraph". We now show how 
many paragraphs have been been inserted into the Piecetable in the status bar.

All this shows the user that we're doing useful things while importing 
big douments.

Windows and OSX developers should implement this for their platforms.



git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@29714 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6