Allow user to set default codepage.
For fonts which are no font substitutes it is impossible to say what the
underlying codepage is, i.e. which codepage was used when the document
was written.

By passing an importer argument, namely '-i mswrite-codepage:CP1250'
(CP1250 in this example), it is possible to set the favored codepage for
these fonts now.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30103 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6