Imported a great deal of the code from the GCT branch. This covers everything that is not in plugins/opendocument. The remains are in odf-2011-track-changes-git-svn/git-diff-2011-oct-3-cleaned.patch.
Some code is added here but enclosed in #ifdef ODF_GCT guards to disable it until it is required for the ODF plugin.
Handy things like pf_Frag::getNextStrux() and tryDownCastStrux() are now available. These should help cleanup
code that wants to walk strux and code that had if( dynamic_cast<> ) to find out of a frag is a specific strux type.
Note that Piecetable::_findLastStruxOfType() now has a stop condition and directly returns the downcast object type
to help avoid such cases in calling code.



git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30356 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6