Remove the text column definition from the .cpp file.
Put it into the .ui files instead.

Moreover, a GtkComboBox with entry is only defined for GTK3 builds,
so it is wrong to set entry-text-column starting with version 2.24.
The GTK2 build uses a GtkComboBoxEntry which requires text-column to
be set.

This fixes a crash with pre-2.24 GTK builds when selecting an entry
from the combo boxes.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30979 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6