Localize style names in Styles dialog.
In order to do so, change the GtkListStore to hold both the original and
the localized names. Display the localized names and internally used the
original ones.

Additionally, fix a problem with "basedon" and "followedby" property
values caused by earlier localization of DLG_Styles_DefNone and
DLG_Styles_DefCurrent. Now the English style names will be used again to
set the values.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@31868 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6